% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_confusion_matrix_plot.R
\name{create_confusion_matrix_plot}
\alias{create_confusion_matrix_plot}
\title{Creates a confusion matrix plot to compare predicted classes with actual classes}
\usage{
create_confusion_matrix_plot(actual, predicted_classes)
}
\arguments{
\item{actual}{A vector of the ground truth classes}

\item{predicted_classes}{A vector of the predicted classes}
}
\value{
A ggplot object of a confusion matrix
}
\description{
This function creates a confusion matrix plot when it is given two vectors, one of the predicted classes and
one of the actual ground truth.
}
\examples{
actual <- c(0, 1, 1, 1)
predicted_classses <- c(0, 0, 1, 1)
create_confusion_matrix_plot(actual, predicted_classes)
}
