% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dummy_col.R
\name{make_dummy_col}
\alias{make_dummy_col}
\title{Transforms a column into dummy variables 1 or 0}
\usage{
make_dummy_col(data_frame, col_name, values)
}
\arguments{
\item{data_frame}{A data frame or data frame extension (e.g. a tibble).}

\item{col_name}{The name of the column (a single string) to check to convert its values to binary}

\item{values}{A vector of values to be treated as one c(...) or a singular value of any type except a dataframe or tibble}
}
\value{
A data frame with a one column as binary variables
}
\description{
Converts one column in a dataframe into dummy variables
}
\examples{
df <- data.frame(
  marital_status = c("Married", "Widowed", "Single", "Widowed")
)
make_dummy_col(df, "marital_status", "Widowed")


}
