% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correl_heatmap.R
\name{plot_correlation_heatmap}
\alias{plot_correlation_heatmap}
\title{Plot a correlation heatmap of numeric variables}
\usage{
plot_correlation_heatmap(data, title = "Correlation Heatmap")
}
\arguments{
\item{data}{A data frame containing numeric columns to correlate}

\item{title}{A string for the plot title}
}
\value{
A ggplot object of the correlation heatmap
}
\description{
Plot a correlation heatmap of numeric variables
}
\examples{
plot_correlation_heatmap(mtcars, "Correlation Heatmap of mtcars")
}
