% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_logistic_model.R
\name{train_logistic_model}
\alias{train_logistic_model}
\title{Train a logistic regression model}
\usage{
train_logistic_model(df, formula)
}
\arguments{
\item{df}{A data frame containing the dataset}

\item{formula}{A formula specifying the model}
}
\value{
A trained glm model object
}
\description{
This function trains a logistic regression model using a given dataset
and formula.
}
\examples{
df <- data.frame(x = c(1,2,3), y = c(0,1,0))
train_logistic_model(df, y ~ x)
}
