% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_confusion_matrix_plot.R
\name{create_confusion_matrix_plot}
\alias{create_confusion_matrix_plot}
\title{Creates a confusion matrix plot from a confusion matrix table}
\usage{
create_confusion_matrix_plot(cm_table)
}
\arguments{
\item{cm_table}{A table or data frame representing the confusion matrix
(e.g., cm$table from caret::confusionMatrix)}
}
\value{
A ggplot object of a confusion matrix
}
\description{
This function takes a confusion matrix table (e.g., from caret::confusionMatrix)
and returns a ggplot visualization.
}
\examples{
library(caret)
actual <- factor(c(0, 1, 1, 1))
predicted <- factor(c(0, 0, 1, 1))
cm <- confusionMatrix(predicted, actual)
create_confusion_matrix_plot(cm$table)
}
